backend/pyproject.toml
[tool.poetry]
name = "naviera-backend"
version = "0.1.0"
description = "Naviera backend (FastAPI API service)"
authors = [
    "Prashant Ranjan <prashantnk2001@gmail.com>"
]
readme = "README.md"
packages = [{include = "app"}] 

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.119.0,<0.120.0"
uvicorn = {extras = ["standard"], version = ">=0.37.0,<0.38.0"}
sqlmodel = ">=0.0.27,<0.0.28"
asyncpg = ">=0.30.0,<0.31.0"
alembic = ">=1.17.0,<2.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
python-jose = ">=3.5.0,<4.0.0"
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
httpx = ">=0.28.1,<0.29.0"
pydantic-settings = "^2.11.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.2,<9.0.0"
pytest-asyncio = ">=1.2.0,<2.0.0"
black = ">=25.9.0,<26.0.0"
ruff = ">=0.14.0,<0.15.0"
isort = ">=7.0.0,<8.0.0"
pre-commit = ">=4.3.0,<5.0.0"
supabase = "^2.22.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
seed = "scripts.seed:main_wrapper"